+++ DawnLib.Dusk/src/API/Definitions/Items/DuskItemDefinition.cs
@@ -25,7 +25,7 @@ public class DuskItemDefinition : DuskContentDefinition<DawnItemInfo>
     public DuskPricingStrategy? PricingStrategy { get; private set; }
 
     [field: Space(10)]
-    [field: Header("Configs | Spawn Weights | Format: <Namespace>:<Key>=<Operation><Value>, i.e. magic_wesleys_mod:trite=+20")]
+    [field: Header("Configs | Spawn Weights")]
     [field: SerializeField]
     public List<NamespacedConfigWeight> MoonSpawnWeightsConfig { get; private set; } = new();
     [field: SerializeField]